♻️ moves document form to a separate web component
🌐 updates translations
