+++ lab/documents/assets/js/pages/document.js
@@ -1,13 +1,13 @@
 "use strict";
 
 import { FileTable } from "../../../../assets/js/components/file-table.js";
-import { FileUploadForm } from "../../../../assets/js/components/file-upload-form.js";
+import { DocumentUploadForm } from "../components/document-upload-form.js";
 import { FileManager } from "../../../../assets/js/file-manager.js";
 import { DocumentPresignedUrlService } from "../presigned-url-service.js";
 import { S3Service } from "../../../../assets/js/s3-service.js";
 
 FileTable.init();
-FileUploadForm.init();
+DocumentUploadForm.init();
 
 const projectId = parseInt(document.URL.split("/").reverse()[1]);
