+++ lab/templates/components/file_upload_form.html
@@ -1,5 +1,6 @@
 {% load i18n %}
-<form is="file-upload-form" {% if component.attrs %}{% include "attrs.html" with attrs=component.attrs %}{% endif %}>
+<form is="{% if component.component_name %}{{ component.component_name }}{% else %}file-upload-form{% endif %}" 
+      {% if component.attrs %}{% include "attrs.html" with attrs=component.attrs %}{% endif %}>
     <div class="fr-mb-4w">
         <div class="fr-upload-group">
             <label class="fr-label" for="file-upload">{% translate 'Add files' %}
