+++ lab/templates/documents/project_documents.html
@@ -26,7 +26,7 @@
 
 <div id="content-main">
     <ul class="messagelist"></ul>
-    {% include "components/file_upload_form.html" with component=file_upload_form %}
+    {% include "components/file_upload_form.html" with component=document_upload_form %}
     <div class="documents-wrapper">
         <h2>{% translate 'Documents' %}</h2>
         <div>
